refactor: move editor related components to dedicated folder
