+++ src/core/components/common/editor/CoursSaveButton.tsx
@@ -1,7 +1,7 @@
 import { Button } from "@/core/components/ui/button";
 import React from "react";
-import useUpdateCoursBody from "../../application/usecases/services/useUpdateCoursBody";
-import { Cours } from "../../domain/entities/cours-schemas";
+import useUpdateCoursBody from "../../../../features/cours-sequence/application/usecases/services/useUpdateCoursBody";
+import { Cours } from "../../../../features/cours-sequence/domain/entities/cours-schemas";
 import { useCurrentEditor } from "@tiptap/react";
 
 function CoursSaveButton({ cours, userId }: { cours: Cours; userId: string }) {
