+++ src/core/components/common/editor/EditorChild.tsx
@@ -1,15 +1,13 @@
 "use clent";
 import { Button } from "@/core/components/ui/button";
 import { useCurrentEditor } from "@tiptap/react";
-import useUpdateCoursBody from "../../application/usecases/services/useUpdateCoursBody";
-import { Cours } from "../../domain/entities/cours-schemas";
+import useUpdateCoursBody from "../../../../features/cours-sequence/application/usecases/services/useUpdateCoursBody";
+import { Cours } from "../../../../features/cours-sequence/domain/entities/cours-schemas";
 
 export default function EditorChild({
- 
   children,
 }: {
   children: React.ReactNode;
-  
 }) {
   const { editor } = useCurrentEditor();
