+++ src/core/components/common/editor/EditorProvider.tsx
