+++ src/core/components/common/editor/MenuBar.tsx
