+++ src/core/components/common/editor/MenuBold.tsx
