+++ src/core/components/common/editor/MenuImage.tsx
