+++ src/core/components/common/editor/MenuTextColor.tsx
